Explicitly set required services in devtest
We currently pull in the required OpenStack services for a cloud implicitly
via boot-stack's element-deps in DIB.  Instead, we can state these as explicit
lists in the various stages of devtest. This is a first step toward allowing
services to be split out of the controller nodes and onto their own machines.
Later, we can expose these lists to boot-stack and other components of the
deployment via Heat.

Change-Id: Ia6b1d344b1d3ee0f226a8b6db0700e79feb0d7e9
Partial-bug: #1340467