Document NeutronControlPlaneID
NeutronControlPlaneID must be passed as a parameter when creating the
overcloud. Otherwise the stack create will fail with: ERROR: Either
network or network_id should be specified, as the OS::Neutron::Port type
requires network to be set to a non-empty string.

This does not show up in the CI because the parameter is set when
devtest was run as a script, but it needs to be documented as well for
those following along from the docs.

Change-Id: Ie115a042ea409380c27b2c4d770a223ce74d8fad