Revert "Set default OVERCLOUD_CONTROLSCALE to 3"
This reverts commit f6dd7fefca2ac7a9ac1ec6394a7e5763b8104b42.

Having our default set to one is reasonable for developers (who
generally have more limited resources).

This does not prevent CI or developers who have extra resources
from testing multi-controller deployments however.

We spoke about this in Atlanta and there was not concensus
that setting the "Dev" test scripts default to 3 was a good
idea.

Change-Id: Ib362bf3b68f95500d961c6fdcc97619f9408dc8a