Improve warning when TRIPLEO_ROOT is defaulted
The warning when TRIPLEO_ROOT was left to default to ~/.cache/tripleo
did not make it clear that other environment variables are also given
defaults based on that value. So just resetting TRIPLEO_ROOT would
not be enough to give you a working system.

This patch just makes the warning a bit clearer.

Change-Id: Idb1091c252ac83f728cb5734884c5992dd0a1754