Allow user to keep virsh instances alive
Any call to devtest_testenv.sh attempts to remove all virsh instances
defined in that test environment. A user might want to configure a set
of VMs manually for use in the undercloud and overcloud later on, but
still take advantage of the functionality provided by _testenv.sh

This patch adds an optional parameter to devtest_testenv which, if
supplied, skips the destruction of UC and OC VMs. A user who does this
will have to provide their own nodes.json which describes the VMs they
have created. The seed VM is still destroyed and recreated as it is
unlikely to need any customisation. This makes the VM-deploy and
BM-deploy more aligned, in that you can have a set of predefined nodes
for your tripleo system and boot a generic seed to deploy onto them.

Part of bp targeted-node-deployment

Change-Id: If7679d74108bc7064df268c240df5fff6a76d69b