Add first draft of HACKING.rst
Consistent style is a highly-desirable feature of a codebase. There
seems to be a recognition of this in review comments, but there is no
guidance for what style should be consistently used.

This patch attempts to document our conventions. It is presented as a
file in the working tree rather than a spec. This is modelled after the
approach taken in the nova project and others. I expect other TripleO
repos to link back to this file rather than contain their own hacking
rules.

In order to ease the creation of this doc, I have tried to only include
uncontentious guidelines. Discussion on additional items will be most
productive on follow-on commits.

Change-Id: Ice8c36f8d10b289aeebab01a19c6035a1be85422