Add first-draft of HACKING.rst
Consistent style is a highly-desirable feature of a codebase. There
seems to be a recognition of this in review comments, but there is no
guidance for what style should be consistently used.

This patch attempts to document our conventions.  It is presented as a
file in the working tree rather than a spec as I believe it should be
mutable to allow clarifications and corrections. This is modelled after
the approach taken in the nova project.

Change-Id: Ice8c36f8d10b289aeebab01a19c6035a1be85422