Replace extract-docs.awk with extract-docs.py
extract-docs.py has several nice features over extract-docs.awk. The
obvious advantage is that it's easier for python developers to read than
the awk script (not that the awk script was particularly hard)

More importantly, this script has the property that it preserves line
numbers - so when you get a failure on line 206 of devtest.rst, that
means the error is on line 206 of devtest.sh, rather than being on some
random other line number.

I'd like to see this script integrated into Sphinx as a dedicated
pre-processor, but in the meantime this should help us make errors
building docs easier to read.

Change-Id: Ibb6f943e3f6b30f1bdbc0b8d7b73f7b32f75fc6a