Don't create extra nodes if nodes file is provided
Unless NODE_CNT is pre-set, we default to creating 15 new nodes, even if
--nodes is used to point at a set of existing nodes. This requires
prebuilt environments to be created by passing --nodes *AND* setting
NODE_CNT to -1

This change makes it easier to use a prebuilt environment; if --nodes is
supplied, NODE_CNT will default to 0 - but an external setting will
still be observed.

It's possible that this change is rendered moot by other work that is
happening to seperate the bm network and node manipulation from the
create-a-test-env portions of the script.

Change-Id: I0291ba5b83ab0ffce7ea7f9e956ee1c1db88bc5b