Exposed fixed VIP address to Heat ControlFixedIPs parameter
When an SSL certificate is used to secure the overcloud,
its common name will dictate the IP address that should be
used for the cluster VIP managed as a Neutron port.

This change recognizes when the PUBLIC_API_URL is set
(via a shell argument to devtest_overcloud.sh) and builds
a JSON string to be provided to Heat via the ControlFixedIPs
parameter.

Change-Id: Icb23248b7bd5347a5cac78403a702b321316bc9c