Whitespace cleanup
In an unrelated changed I noticed that some of our scripts have lines
with trailing whitespace after a ##, to avoid unexpected indentation
errors from sphinx.

For the most part this is unnecessary, a pure blank line works just as
well (and doesn't have trailing whitespace), so I've removed those
lines.

Change-Id: I07be1301c95c5c994834858c0a20731dc92eb49d