Detect rhel7 properly
set-os-type was setting all versions of RHEL to use the "rhel"
element.  This won't work for RHEL 7, so we need to extend the
check to handle that correctly.

Note that the full version string in redhat-release is:
Red Hat Enterprise Linux Server release 7.0 (Maipo)
so this should handle point updates too as long as they don't
change the format.

Change-Id: I730dce701c94799f0efcc3ae8096ce9930d80170