Add USE_IRONIC_NOVA_DRIVER option
Adds a new option to switch on the ability to use the new
Ironic nova driver location (in the Nova driver tree).

In order for this to work we need to expose some new
heat parameters here:

 Ib48a6b6d8a6cff157bdf3948bd0330e9b29dd46a (https://review.openstack.org/#/c/119472/)

We also need to update the nova.conf template to support the setting
scheduler_host_manager (previously hard coded):

 Id4ca7f90cd85129727e65d0d09f274f3a00d4ebe (https://review.openstack.org/#/c/119633/)

The change is backwards compatible and defaults to using the old Ironic
Nova compute driver for now as to not break anything.

Change-Id: I1138ac7c0000db41eceee9420cbd54c9cacf922d