Add a pep8 testenv
To allow us to start thinking about using bashate for our scripts, I've
added a pep8 testenv, which runs bashate and only bashate. The confusing
name is because we expect to add flake8 to this environment later, and
we want the name to be consistent with other projects.

Currently I'm seeing 419 errors; largely trailing whitespace,
indentation, and function calls using the 'wrong' format.

Change-Id: Ibc04e00d4278e3656e1e5981a6c1b646cd61f39f