Scale the wait_for_hypervisor_stats linearly
When we're building small clouds, or  baremetal VMs, waiting 60 seconds
for them all to come up seems reasonable. On larger clouds on physical
hardware, this can be a less useful timeframe.

Rather than setting a larger default (and still be too short for some
clouds, and be much longer than a small cloud needs), this patch changes
from 60 loops of 1 second each; to 60 loops of
one-second-per-expected-node each.

Change-Id: Ic3481cb1a1d104c02f4407955f75c626a8366ee0