Adds vcpus and memory parameters
Adds vcpus and memory parameters to the wait_for_hypervisor_stats
script. Not all nodes will have identical hardware characteristics,
especially in a baremetal environment, so this script needs to be a bit
more flexible. Will follow up with a subsequent patch that makes use of
this functionality in devtest.

Maintains backwards compatibility.

Change-Id: I43d72cc0d2c1ded70af28d9861aa6e28838795bf