Allow the PS4 mangling to be turned off
The PS4 mangling that devtest.sh does is very useful in showing
exactly where an issue with a devtest run is going wrong. However,
for users that have a TRIPLEO_ROOT set, this turns the PS4 prompt
into a mass of 70 lines, which harms debugging. Check a variable
is not 1 before doing so.

Change-Id: Ie55e714d50622ac38f36f67a26af5b138c1d2f48