Support passing x86_64 to create-nodes
x86_64 is the canonicalized arch for amd64 in Nova[1]. I think it makes
sense to make create-nodes (and thus our UI) a bit friendlier by making
it as an accepted arch value. This will allow for the baremetal flavor
to be created with x86_64 set as cpu_arch in the extra_specs so that if
folks happened to register nodes with x86_64 instead of amd64 in Ironic,
the flavor matching will work.

[1] http://git.openstack.org/cgit/openstack/nova/tree/nova/compute/arch.py#n156

Change-Id: I0f207a5b190bc2df86bc58fc62abc5ac6a93d73f