Remove variable case requirement
Although the requirement disambiguates exported from non-exported vars,
which is useful, few if any of our existing scripts apply it. This makes
it difficult to enforce on new commits. I'd rather not have HACKING
rules which are unenforced.

I'd not object to this being reintroduced along with a cleanup to
existing scripts. I do think that'll be impractical but if someone feels
strongly they're welcome to do the work.

Change-Id: Iafde3c94caeb81f5daf0b97bffc61155fcadee76