Check for firewalld before starting seed
Now that we are watching for a completion signal from the seed, we
need to make sure that a firewall isn't going to block the
connection.  On Fedora, firewalld is enabled by default on a
minimal install so we need to check for it and stop devtest with
a message about how to fix the problem.  The port is not
automatically opened to avoid silently poking holes in the user's
firewall, which might not be appreciated in strict environments.

Change-Id: I98d5ef17960c7cb963d1934c609d902a4492570d