Add support for specifying flavors
Currently, the undercloud, overcloud control and overcloud compute
flavors are hardcoded to 'baremetal', which makes it harder for
operators to decide what gets deployed where when they have a
heterogeneous environment. Allow the flavor to be deployed onto to
be easily overridden, leaving the default as 'baremetal' for
backwards compatibility.

Change-Id: I1b653020acf4b5acb510ba4a092f854815e36a87