Switch to new setup-flavors
Switch to the new setup-flavors utility provided by os-cloud-config.
This doesn't remove any existing script from this repository, but
replaces functionality in setup-baremetal.

This also removes the hardcoded 'baremetal' as the flavor name for
the undercloud and overcloud scripts, since the scripts will now
programmatically work out the flavor name, and raise an error if
there is more than one.

Change-Id: Ib150de715d6f09c57f920fa3755675718a2e76d1