Fix call to show_options if no BUILD_COMMAND set
Fixes an issue where if you called acquire-image with less
that 2 ARGS it would fail without printing any error message.
This is because the IMAGE_NAME and/or BUILD_COMMAND variables
were obtained from $1 which may or may not be set.

To work around this we do a simple error check on the number of args
before setting (and checking) these variables. This makes it
so acquire-image properly prints its usuage message if args aren't
present.

Change-Id: Iaa1209923fd5259a153312df65519d31f1d929a1