acquire-image: respect local image caching w/ -c
If a .qcow2 image file exists and the -c option is
in used then don't rebuild the image. This updates
acquire-image so that it matches the previous
behavior of the devtest_* scripts where local
.qcow2 image files worked file when USE_CACHE
was set.

This patch also removes a confusing statement in the help
message that states that a .meta metadata file is what
determines if an image is present.

Change-Id: I96243d64aafe5fd964952e348eb7f31fa5cab4cb
Closes-bug: #1392732