Add support for building Ceph nodes and configure Cinder to use it
Introduces support for CEPHSTORAGESCALE and CEPHSTORAGE_FLAVOR,
the first meant to define the number of Ceph OSDs to deploy, the
second the nova flavor to use for the Ceph OSD nodes.

Also adds the logic which enables the usage of Ceph elements and the
configuration of the Cinder backend basing on the value of
CONTROLSCALE (if >1 that is).

Change-Id: I6afadbcf7b17b11ec1a78e84cb1a4732d2646e2f
Depends-On: I9d5f9da4e411acfb55658a104e061a97fc82be1a
Depends-On: I954861536c41b2a7e6cbd86a0f0b55004eed4c70