Add --ha param to devtest
We currently provide control to whether a deployment is HA by tuning the
OVERCLOUD_CONTROLSCALE/OVERCLOUD_COMPUTESCALE environment variables.
This is a pretty poor experience for users in that we should not support
arbitrary numbers of control nodes but we still require users to specify
the control node count they desire.

A much better interface is for the user to specify whether the
deployment should be HA or not, and allow the user to specify an
arbitrary number of compute nodes. This helps us move in the direction
of having a few simple devtest commands to fit a few primary use cases:
minimal deployments for dev/testing, and production deployments.

Change-Id: I08df10b73e6f266a31019c968b77a03e1f696009