Make wait_for use getopt and add walltime support
We currently use wait_for which does not account for time spent blocking
during COMMAND. This leads to issues where it is hard to calculate time
we want to spend waiting for something. Creating a new tool which blocks
based on wall time.

Also adding progress '.''s to help debugging.

Closes-Bug: #1407812

Change-Id: Icdc626ef8075fbd2f9e7cb7c011a12351c815e09