Make the default for NODE_CNT less arbitrary
Reduce the number of nodes created on a virt host to 10 so that less
disk space is used. This makes a difference when devtest is run several
times on the same virt host as each time ironic will pick a different VM
on which to deploy nodes too, eventually the qcow being used in all the
defined domains will be more then just a blank image.

Change-Id: I442e2c5f36a5d40052285977303ce8fba57f94bf