Switch USE_MERGEPY default to 0
Older templates which rely on merge.py are deprecated; this change
makes devtest use by default the newer templates which use the
equivalent functionalities provided natively by the HOT language.

The --no-mergepy option is not removed as external scripts might
still rely on it. The older templates can still be used by setting
USE_MERGEPY=1 in the devtest environment.

Change-Id: Ibafd7325ff1310e029074996e8ccb31981cf1f1e