build-images: drive DIB via YAML config file
Adds a new build-images python script which can be used to build
multiple DIB images at once using a YAML (or JSON) config file.

The YAML configuration format looks like this:
  disk_images:
    -
       imagename: overcloud-compute
       arch: amd64
       type: qcow2
       elements:
         - overcloud-compute
       packages:
         - vim
       options:

The motivation for this script is to be able to use it
as a replacement for the growing number of environment
variables required to get devtest_overcloud.sh working
in all environments.

Change-Id: I5087b178253805379778ab19f093483d56b5a811