Wire in load-images
Updates devtest_overcloud.sh so that it makes use of load-images
to load images in a more flexible format via the load-images utility.
Images created with devtest_overcloud_images.sh will be loaded via
the static overcloud_disk_images.yaml file. If a custom YAML
script is provided to build images that will also be used to
load images as well.

Change-Id: I78c4c532169721674b2ea58ef37d2a4b23693fae