Allow a dev to specify puppet-modules override
Currently it is pretty cumbersome to specify the puppet module one wants
to test since they are put in the image during build.

The idea of this commit is to allow one to specify a list of module s/he
wants to override for a given run

A puppet-modules file ($TRIPLEO_ROOT/puppet-modules) is formatted as follow:

tripleo https://review.openstack.org/stackforge/puppet-tripleo master
refs/changes/58/175958/2
rabbitmq https://github.com/me/puppetlabs-rabbitmq mybranch

Both puppet-tripleo and puppet-rabbitmq will be overiden in the images
at /opt/stack/puppet-modules/ to apply the one that were specified in
this file.

Change-Id: I622fa354b1129f8152895e876eb488c6bbab1e6a