QuintupleO setup environment
Adds setup docs, heat templates and scripts for a quintupleo
environment as an alternative to virt and baremetal.

Many of these scripts were originally derived from
https://github.com/cybertron/tripleo-scripts

quintupleo/README.rst is derived from rdo-management environment
setup, so further refinement is likely required to work with an
image based undercloud.

Co-Authored-By: Ben Nemec <bnemec@redhat.com>

Change-Id: I99847703f551db302b4a258076957af9ba8bcbf7
Implements-Blueprint: tripleo-on-openstack