Allow configurable bridge names and clean all.
Adds two new (related) options to cleanup-env that can
be used when multiple baremetal interfaces are getting
assigned to nodes.

The -b option can be used to specify multiple bridge
names to cleanup-env. This option is useful because
VMs are named based on the bridge name(s). This option
works similarly to the -n option and can be used to
cleanup only specific nodes which match a regex.

The -a option can be used to cleanup everything locally
regardless of name. I found myself wanting this option when
cleaning up an environment created with multiple bridges
on the CLI manually. This avoids having to specify anything,
and just cleans it all.

Change-Id: I94d1bcf4112c1e3f62c14abb7028f7215b6e7662