cleanup-env: Allow configurable bridge names.
Add a -b option that can be used to specify multiple bridge
names to cleanup-env. This option is useful because
create-nodes VMs are named based on the bridge name(s).
This option works similarly to the -n option and can be used to
cleanup only specific nodes which match a regex.

Change-Id: I94d1bcf4112c1e3f62c14abb7028f7215b6e7662