Document needed pins for Puppet
We've been running with these in CI for a while. While it's not ideal to
have to document them here as well, there's no way for someone to try
anything with Puppet because we know they will fail when not using these
pins.

If this spot isn't ideal, I'm open to other ideas about where to commit
this info  but we need to doc this stuff somewhere so we don't have to
tell people who hit issues to "go look at what tripleo-ci is using".

Change-Id: Ic76c974e162f61e117a97376a9ca54b608b51f63