Add docker_images_redhat.yaml build-images file
This patch adds a file to help build the heat docker "agent"
base image. We intend to use an image like this to help provision
a docker atomic host for TripleO. It should allow us to
run os-collect-config, various plugins, among other things.

Building this image with DIB is a really handy way to tap
into existing elements for TripleO which already support
the things we need. The downside is that the generated
image is going to be a base image... so there may still
be a case for building a similar image with Dockerfile's
as well. Still... having the the capability to work
with existing elements should significantly speed
up development because we can make use of more robust
elements like source-repositories, etc.

Because of how the docker base element works (and isn't a full
operating system) have had to define elements like yum, and
redhat-common within this config file. As such it is Red Hat
specific.

We also exclude the 'base' element to keep things a bit smaller.

Change-Id: I195c0ef8b844a5fd28a1f4ba465d975787775734