jq 1.5 support
According the JSON spec, the \n character should be encoded. That's a
policy enforced by jq 1.5.

Because of that, it's not possible anymore the generate a JSON file with
the content of the private key without dropping the \n.

Without this patch, we get the following error:
parse error: Invalid string: control characters from U+0000 through
U+001F must be escaped at line 32, column 30

Change-Id: Id4a956443ff8c99f76a5778ed5c3a53367c8e408