Replace the hard-code home directory with a tilde.
Hard-coded user`s home directory will cause problems, if a customized base directory is specified by the HOME variable into the /etc/default/useradd.
For instance,
  $cat /etc/default/useradd
     #HOME=/home
     HOME=/tmp/

Then, the directory "/tmp/" (not "/home/") will be the default prefix for the home directory of the new user which is created by "useradd" with "--create-home" option.

Change-Id: I4e343955a187195652b1ecca5982869bd5c8133c