Mark this repo as deprecated
The readme in incubator is horribly out of date, and I hear it
still comes up above the current docs in a Google search for TripleO
(need better SEO!).  Since I believe all of the still relevant
details from the incubator readme are now in the tripleo-docs repo,
let's just kill this one and replace it with a link to the new docs.

Change-Id: I3dbc263a012ea7269db42de66b2ffcc0003fd27d